--- components/hooks/useWindowWidth.js
@@ -1,23 +1,0 @@
-import { useState, useEffect } from 'react';
-
-function useWindowWidth() {
-  const [isSmallerDevice, setIsSmallerDevice] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      const width = window.innerWidth;
-      setIsSmallerDevice(width < 500);
-    };
-
-    handleResize();
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
-
-  return { isSmallerDevice };
-}
-
-export default useWindowWidth;
