+++ pages/_app.js
@@ -1,9 +1,11 @@
 import React from 'react';
-
+import { WindowWidthProvider } from '../components/hooks/windowWidthContext';
 const App = ({ Component, pageProps }) => (
-  <React.Fragment>
-    <Component {...pageProps} />
-  </React.Fragment>
+  <WindowWidthProvider>
+    <React.Fragment>
+      <Component {...pageProps} />
+    </React.Fragment>
+  </WindowWidthProvider>
 );
 
 export default App;
